SplunkADXForwarder/requirements.txt
azure-kusto-data == 4.1.4
azure-kusto-ingest == 4.1.4
PyYAML == 6.0
pandas == 2.0.1